requirements.txt
python-dateutil
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-environ
requests